requirements.txt
aiofiles==24.1.0
aioredis==1.3.1
aiosqlite==0.17.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apache-libcloud==3.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.3
asyncpg==0.29.0
Authlib==1.3.2
babel==2.16.0
bcrypt==4.1.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cryptography==43.0.0
databases==0.9.0
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.2
fastadmin==0.2.9
fastapi==0.111.1
fastapi-admin==1.0.4
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Naked==0.1.32
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
sqlalchemy-file==0.6.0
starlette==0.37.2
time-machine==2.15.0
tortoise-orm==0.21.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
WTForms==3.1.2
